// src/types/copySettings.ts
export interface CopySettings {
  mainTabs: boolean;
  subTabs: boolean;
  buttons: boolean;
  menus: boolean;
  cards: boolean;
  badges: boolean;
  textControls: boolean;
  imageControls: boolean;
  analytics: boolean;
  socialPosts: boolean;
  notifications: boolean;
  tooltips: boolean;
  popups: boolean;
  sliders: boolean;
  forms: boolean;
  opacity: number;
}